ws10 complete. will probably need changes for sign off.
Wrote a	runMenuWithForLoop()	method	for  the	Application	class	that	5
times	(using	a	for	loop)	allows	the	user 	to	do	one	of:
   a. change	the	current	time;
   b. change	the	alarm	time;
   c. set	or	unset	the	alarm.
Made	sure	 this uses a	command	line	menu	with	a	switch	statement.
Wrote	a	new	version	of	the	method	called	runMenuWithWhileLoop()
that	uses a	do	..	while loop.
Changed the	toString()	method	in	the	Time	class	so	that	it	has	the	code:
return	String.format("%02d:%02d",	hour,	minutes);
Add checks so that the time entered is actually real time on the 24 hour clock

Signed-off-by: chm60 <chm60@aber.ac.uk>